Move ec2 model utilities to separate modules.
Split up utilities by cloud and by account vs job.
